rename mongo stats functions and combine files
