+++ src/api/main.py
@@ -46,7 +46,6 @@ from api.views.cycle_views import (
     CycleStatsView,
     CyclesView,
     CycleView,
-    MongoCycleStatsView,
 )
 from api.views.db_views import DatabaseManagementView
 from api.views.failed_email_views import FailedEmailsView, FailedEmailView
@@ -113,7 +112,6 @@ rules = [
     ("/cycle/<cycle_id>/", CycleView),
     ("/cycle/<cycle_id>/manual_reports/", CycleManualReportsView),
     ("/cycle/<cycle_id>/stats/", CycleStatsView),
-    ("/cycle/<cycle_id>/mongostats/", MongoCycleStatsView),
     ("/cycle/<cycle_id>/reports/<report_type>/", ReportHtmlView),
     ("/cycle/<cycle_id>/reports/<report_type>/pdf/", ReportPdfView),
     ("/cycle/<cycle_id>/reports/<report_type>/email/", ReportEmailView),
